requirements.txt
Django==3.1.3
django-crispy-forms==1.11.1
django-debug-toolbar==3.1.1
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.2
